Refactor notifications.
