+++ App.js
@@ -41,6 +41,8 @@ import Group from "./components/Groups/Group";
 import GroupSettings from "./components/Groups/GroupSettings";
 
 import Card from "./components/components/Card";
+import List from "./components/components/List";
+import ListElement from "./components/components/ListElement";
 
 import * as firebase from "firebase";
 import config from "./config";
@@ -179,6 +181,16 @@ export default class App extends React.Component {
     };
   }
 
+  date = new Date();
+
+  dateStr = `${this.date.getFullYear()}-${
+    this.date.getMonth() + 1 < 10
+      ? `0${this.date.getMonth() + 1}`
+      : this.date.getMonth() + 1
+  }-${
+    this.date.getDate() < 10 ? `0${this.date.getDate()}` : this.date.getDate()
+  }`;
+
   busynessMessage = [
     "No one else is here!",
     "Still easy to find a table.",
@@ -306,7 +318,11 @@ export default class App extends React.Component {
       await this.updateProfile(() => console.log("profile loaded"));
       await this.updateFriends(() => console.log("friends loaded"));
       await this.updateGroups(() => console.log("groups loaded"));
-      await this.updateRatings();
+      await this.resetNotifications(
+        () => console.log("notifications reset"),
+        true
+      );
+      await this.updateRatings(() => console.log("ratings loaded"));
       // TODO: Get the user's ratings here
 
       await this._storeData(`user`, JSON.stringify(this.state.user));
@@ -332,8 +348,7 @@ export default class App extends React.Component {
         user: {
           ...this.state.user,
           ...data,
-          friends: [],
-          groups: []
+          notifications: this.state.user.notifications
         }
       });
     });
@@ -354,12 +369,12 @@ export default class App extends React.Component {
   };
 
   updateFriends = async callback => {
-    /*await this.setState({
+    await this.setState({
       user: {
         ...this.state.user,
         friends: []
       }
-    });*/
+    });
 
     await this.fetchFriends(
       this.state.user.userHandle,
@@ -373,12 +388,12 @@ export default class App extends React.Component {
   };
 
   updateGroups = async callback => {
-    /*await this.setState({
+    await this.setState({
       user: {
         ...this.state.user,
         groups: []
       }
-    });*/
+    });
 
     await this.fetchGroups(this.state.user.userHandle, async data => {
       await data.forEach(
@@ -439,7 +454,7 @@ export default class App extends React.Component {
     }
   };
 
-  updateRatings = async () => {
+  updateRatings = async callback => {
     try {
       let data = await fetch(
         "https://us-central1-courtsort-e1100.cloudfunctions.net/getUserDishRatings",
@@ -455,28 +470,15 @@ export default class App extends React.Component {
         }
       );
       let ratingData = await JSON.parse(data._bodyText);
-      await this.setState({
-        user: { ...this.state.user, ratings: ratingData }
-      });
+      await this.setState(
+        {
+          user: { ...this.state.user, ratings: ratingData }
+        },
+        callback
+      );
     } catch (error) {
       console.error(`getUserDishRatings : ${error}`);
     }
-  };
-
-  checkOut = callback => {
-    Alert.alert("Check Out", `Would you like to check out?`, [
-      {
-        text: "Cancel"
-      },
-      {
-        text: "No"
-      },
-      {
-        text: "Yes",
-        onPress: () =>
-          this.checkOutOfDiningCourt(() => this.setStatus(0, callback))
-      }
-    ]);
   };
 
   updateDietaryRestrictions = async restrictions => {
@@ -503,13 +505,29 @@ export default class App extends React.Component {
     ).catch(error => console.error(`setDietaryRestrictions : ${error}`));
   };
 
+  checkOut = callback => {
+    Alert.alert("Check Out", `Would you like to check out?`, [
+      {
+        text: "Cancel"
+      },
+      {
+        text: "No"
+      },
+      {
+        text: "Yes",
+        onPress: () =>
+          this.checkOutOfDiningCourt(() => this.setStatus(0, callback))
+      }
+    ]);
+  };
+
   checkIn = (courtId, callback) => {
     Alert.alert(
       "Check In",
       `What status would you like to have while you eat at ${courtId}?`,
       [
         {
-          text: this.statusMessage[0],
+          text: "Cancel",
           onPress: () =>
             this.checkOutOfDiningCourt(() => this.setStatus(0, callback))
         },
@@ -536,7 +554,7 @@ export default class App extends React.Component {
 
     Alert.alert("Change Status", `What status would you like to have?`, [
       {
-        text: this.statusMessage[0],
+        text: "Cancel",
         onPress: () =>
           this.checkOutOfDiningCourt(() => this.setStatus(0, callback))
       },
@@ -553,7 +571,6 @@ export default class App extends React.Component {
     ]);
   };
 
-  // status is an integer.
   setStatus = (status, callback) => {
     this.setState({ user: { ...this.state.user, status: status } });
 
@@ -660,7 +677,7 @@ export default class App extends React.Component {
         ]
       };
 
-      render() {
+      render = () => {
         return (
           <Overlay
             overlayStyle={{ padding: 0 }}
@@ -708,7 +725,7 @@ export default class App extends React.Component {
             </Card>
           </Overlay>
         );
-      }
+      };
     }
 
     return <AlertPicker />;
@@ -726,7 +743,7 @@ export default class App extends React.Component {
         options: busynessMessage
       };
 
-      render() {
+      render = () => {
         return (
           <Overlay
             overlayStyle={{ padding: 0 }}
@@ -762,7 +779,7 @@ export default class App extends React.Component {
             </Card>
           </Overlay>
         );
-      }
+      };
     }
 
     return <BusynessPicker />;
@@ -949,15 +966,6 @@ export default class App extends React.Component {
       .catch(error => console.error(`getDiningCourtRatings: ${error}`));
   };
 
-  updateNotifications = notifications => {
-    this.setState({
-      user: {
-        ...this.state.user,
-        notifications
-      }
-    });
-  };
-
   rateDiningCourt = (diningCourt, rating) => {
     fetch(
       "https://us-central1-courtsort-e1100.cloudfunctions.net/rateDiningCourt",
@@ -981,7 +989,7 @@ export default class App extends React.Component {
       this.setState({ mealsLoaded: true });
       return;
     }
-    let date = new Date();
+    let date = this.date;
     date.setDate(date.getDate() + from);
     const dateStr = `${date.getFullYear()}-${
       date.getMonth() + 1 < 10 ? `0${date.getMonth() + 1}` : date.getMonth() + 1
@@ -1020,7 +1028,374 @@ export default class App extends React.Component {
       .catch(error => console.error(`fetchMeals: ${error}`));
   };
 
-  render() {
+  resetNotifications = async callback => {
+    let n = this.state.user.notifications;
+    this.setState(
+      {
+        user: {
+          ...this.state.user,
+          notifications: []
+        }
+      },
+      () => {
+        let arr = [];
+        n.forEach(list =>
+          list.items.forEach(e => this.parseNotifications(arr, e.type, e))
+        );
+        this.addNotifications(arr, callback);
+      }
+    );
+  };
+
+  updateNotifications = async (callback, noStore) => {
+    this.getNotifications(data => {
+      let arr = [];
+      data.forEach((e, index) => this.parseNotifications(arr, e.type, e.id));
+      this.addNotifications(arr, () => {
+        if (noStore !== true) {
+          console.log("storing ...");
+          console.log(this.state.user.notifications);
+          this._storeData("user", JSON.stringify(this.state.user));
+        }
+        if (callback) callback();
+      });
+    });
+  };
+
+  getNotifications = async callback => {
+    fetch(
+      "https://us-central1-courtsort-e1100.cloudfunctions.net/getNotifications",
+      {
+        method: "POST",
+        headers: {
+          Accept: "application/json",
+          "Content-Type": "application/json"
+        },
+        body: JSON.stringify({
+          userHandle: this.state.user.userHandle
+        })
+      }
+    )
+      .then(data => {
+        this.handleData(`getNotifications`, data, callback);
+      })
+      .catch(error => console.error(`getNotifications: ${error}`));
+  };
+
+  parseNotifications = (notifications, type, id, callback) => {
+    let obj = undefined;
+    console.log(id);
+    switch (type) {
+      case "new friend":
+        obj = this.newFriendNotification(id);
+        break;
+      case "new friend request":
+        obj = this.newFriendRequestNotification(id);
+        break;
+      case "unfriended":
+        obj = this.newUnfriendNotification(id);
+        break;
+      case "blocked":
+        obj = this.newBlockedNotification(id);
+        break;
+      case "invited to group":
+        obj = this.newGroupInvitationNotification(id);
+        break;
+      case "user joined group":
+        obj = this.newGroupJoinNotification(id);
+        break;
+      case "user left group":
+        obj = this.newGroupLeaveNotification(id);
+        break;
+      default:
+        console.error(`invalid notification type: ${type}\nid: ${id}`);
+    }
+
+    if (obj) {
+      notifications.push({
+        ...id,
+        ...obj,
+        type: type,
+        date: this.dateStr
+      });
+    }
+  };
+
+  addNotifications = (arr, callback) => {
+    let n = [];
+    if (this.state.user.notifications)
+      n = this.state.user.notifications.slice();
+
+    arr.forEach(item => this.addNotification(n, item));
+
+    this.setState(
+      {
+        user: {
+          ...this.state.user,
+          notifications: n
+        }
+      },
+      callback
+    );
+  };
+
+  addNotification = (arr, item) => {
+    let list = arr.find(l => l.date === item.date);
+    if (list == undefined) {
+      list = {
+        date: this.dateStr,
+        Name: this.dateStr,
+        items: []
+      };
+      arr.push(list);
+    } else {
+      list.items = list.items.slice();
+    }
+
+    list.items.push(item);
+  };
+
+  removeNotification = id => {
+    let n = this.state.user.notifications.slice();
+    for (let i = 0; i < n.length; i++) {
+      if (n[i].date == id.date) {
+        n[i].items = n[i].items.filter(req => req.Name != id.Name);
+      }
+    }
+
+    this.setState(
+      {
+        user: {
+          ...this.state.user,
+          notifications: n
+        }
+      },
+      () => {
+        console.log("storing ...");
+        this._storeData("user", JSON.stringify(this.state.user));
+      }
+    );
+  };
+
+  dismissNotification = id => {
+    Alert.alert(`Dismiss notification?`, undefined, [
+      {
+        text: "Cancel"
+      },
+      {
+        text: "No"
+      },
+      {
+        text: "Yes",
+        onPress: () => this.removeNotification(id)
+      }
+    ]);
+  };
+
+  newFriendRequestNotification = id => {
+    id.Name = `${id.friendName}  @${
+      id.friendHandle
+    }  would like to become friends`;
+    id.id = id.friendHandle;
+    let obj = { ...id, onPress: () => this.friendAlert(id) };
+    return obj;
+  };
+
+  newFriendNotification = id => {
+    this.updateFriend(id.friendHandle, true);
+
+    id.Name = `${id.friendName} @${
+      id.friendHandle
+    } accepted your friend request.`;
+    let obj = { ...id, onPress: () => this.dismissNotification(id) };
+    return obj;
+  };
+
+  newUnfriendNotification = id => {
+    this.updateFriend(id.friendHandle, false);
+
+    id.Name = `${id.friendName} @${id.friendHandle} unfriended you.`;
+    let obj = { ...id, onPress: () => this.dismissNotification(id) };
+    return obj;
+  };
+
+  newBlockedNotification = id => {
+    this.updateFriend(id.friendHandle, false);
+
+    id.Name = `${id.userName} @${id.userHandle} blocked you.`;
+    let obj = { ...id, onPress: () => this.dismissNotification(id) };
+    return obj;
+  };
+
+  newGroupJoinNotification = id => {
+    this.updateGroup(id.groupID, true);
+
+    id.Name = `@${id.userHandle} joined the group: ${id.groupName}.`;
+    let obj = { ...id, onPress: () => this.dismissNotification(id) };
+    return obj;
+  };
+
+  newGroupLeaveNotification = id => {
+    this.updateGroup(id.groupID, true);
+
+    id.Name = `@${id.userHandle} has left the group: ${id.groupName}.`;
+    let obj = { ...id, onPress: () => this.dismissNotification(id) };
+    return obj;
+  };
+
+  newGroupInvitationNotification = id => {
+    id.Name = `@${id.friendHandle} invited you to join ${id.groupName}`;
+    id.id = id.groupID;
+
+    let obj = { ...id, onPress: () => this.groupAlert(id) };
+    return obj;
+  };
+
+  friendAlert = id => {
+    Alert.alert("Friend Request", `Accept request from @${id.friendHandle}?`, [
+      {
+        text: "Cancel"
+      },
+      {
+        text: "Deny",
+        onPress: () => this.denyFriendRequest(id)
+      },
+      {
+        text: "Accept",
+        onPress: () => this.acceptFriendRequest(id)
+      }
+    ]);
+  };
+
+  groupAlert = id => {
+    Alert.alert(
+      "Group Invite",
+      `Join group ${id.groupName} made by @${id.friendHandle}?`,
+      [
+        {
+          text: "Cancel"
+        },
+        {
+          text: "Deny",
+          onPress: () => this.denyGroupInvitation(id)
+        },
+        {
+          text: "Accept",
+          onPress: () => this.acceptGroupInvitation(id)
+        }
+      ]
+    );
+  };
+
+  acceptFriendRequest = id => {
+    fetch(
+      "https://us-central1-courtsort-e1100.cloudfunctions.net/acceptFriendRequest",
+      {
+        method: "POST",
+        headers: {
+          Accept: "application/json",
+          "Content-Type": "application/json"
+        },
+        body: JSON.stringify({
+          userHandle: this.state.user.userHandle,
+          friendHandle: id.friendHandle
+        })
+      }
+    )
+      .then(data => {
+        try {
+          //JSON.parse(data._bodyText);
+          this.removeNotification(id);
+          this.updateFriend(id.friendHandle, true);
+        } catch (error) {
+          console.error(`acceptFriendRequest: ${error} -- ${data._bodyText}`);
+        }
+      })
+      .catch(error => console.error(`acceptFriendRequest: ${error}`));
+  };
+
+  denyFriendRequest = id => {
+    fetch(
+      "https://us-central1-courtsort-e1100.cloudfunctions.net/denyFriendRequest",
+      {
+        method: "POST",
+        headers: {
+          Accept: "application/json",
+          "Content-Type": "application/json"
+        },
+        body: JSON.stringify({
+          userHandle: this.state.user.userHandle,
+          friendHandle: id.friendHandle
+        })
+      }
+    )
+      .then(data => {
+        try {
+          //JSON.parse(data._bodyText);
+          this.removeNotification(id);
+        } catch (error) {
+          console.error(`denyFriendRequest: ${error}: ${data._bodyText}`);
+        }
+      })
+      .catch(error => console.error(`denyFriendRequest: ${error}`));
+  };
+
+  acceptGroupInvitation = id => {
+    fetch(
+      "https://us-central1-courtsort-e1100.cloudfunctions.net/acceptGroupInvitation",
+      {
+        method: "POST",
+        headers: {
+          Accept: "application/json",
+          "Content-Type": "application/json"
+        },
+        body: JSON.stringify({
+          userHandle: this.state.user.userHandle,
+          friendHandle: id.friendHandle,
+          groupID: id.groupID
+        })
+      }
+    )
+      .then(data => {
+        try {
+          //JSON.parse(data._bodyText);
+          this.removeNotification(id);
+          this.updateGroup(id.groupID, true);
+        } catch (error) {
+          console.error(`acceptGroupInvitation: ${error}- ${data._bodyText}`);
+        }
+      })
+      .catch(error => console.error(`acceptGroupInvitation: ${error}`));
+  };
+
+  denyGroupInvitation = id => {
+    fetch(
+      "https://us-central1-courtsort-e1100.cloudfunctions.net/denyGroupInvitation",
+      {
+        method: "POST",
+        headers: {
+          Accept: "application/json",
+          "Content-Type": "application/json"
+        },
+        body: JSON.stringify({
+          userHandle: this.state.user.userHandle,
+          friendHandle: id.friendHandle,
+          groupID: id.groupID
+        })
+      }
+    )
+      .then(data => {
+        try {
+          //JSON.parse(data._bodyText);
+          this.removeNotification(id);
+        } catch (error) {
+          console.error(`denyGroupInvitation: ${error}- ${data._bodyText}`);
+        }
+      })
+      .catch(error => console.error(`denyGroupInvitation: ${error}`));
+  };
+
+  render = () => {
     /*console.log("render:")
     console.log(this.state.user);
     console.log(":render")*/
@@ -1061,5 +1436,5 @@ export default class App extends React.Component {
       );
     }
     return <Splash />;
-  }
+  };
 }
